rework BeRepositoryBasedId to use 24x40 bits vs. 32x32 bits for repoid/localid. All 64bit ids are now unsigned vs. signed values.

+++ iModelCore/DgnPlatform/DgnCore/Annotations/AnnotationFrameStyle.cpp
@@ -360,7 +360,7 @@ BentleyStatus AnnotationFrameStylePersistence::DecodeFromFlatBuf(AnnotationFrame
         return ERROR;
 
     if (fbStyle->has_id())
-        style.m_id = DgnStyleId(fbStyle->id());
+        style.m_id = DgnStyleId((uint64_t)fbStyle->id());
 
     if (fbStyle->has_name())
         style.m_name = fbStyle->name()->c_str();
